updtae Makefile cd and better free with cmd fail
