+++ src/ast_free.c
@@ -17,27 +17,12 @@ void	free_redirections(t_redir *redirs)
 	}
 }
 
-// void	free_ast_children(t_ast *node)
-// {
-// 	if (node->left)
-// 	{
-// 		free_ast(node->left);
-// 		node->left = NULL;
-// 	}
-// 	if (node->right)
-// 	{
-// 		free_ast(node->right);
-// 		node->right = NULL;
-// 	}
-// }
-
 void	free_ast_cmd(t_ast *node)
 {
 	if (!node || !node->cmd)
 		return ;
 	free_ast_cmd_args(node);
-	if (node->cmd->redirs)
-		free_redirections(node->cmd->redirs);
+
 	if (node->cmd->path)
 		free(node->cmd->path);
 	free(node->cmd);
@@ -48,8 +33,16 @@ void	free_ast(t_ast *node)
 {
 	if (!node)
 		return ;
-	free_ast(node->left);
-	free_ast(node->right);
+	if (node->left)
+	{
+		free_ast(node->left);
+		node->left = NULL;
+	}
+	if (node->right)
+	{
+		free_ast(node->right);
+		node->right = NULL;
+	}
 	free_ast_cmd(node);
 	free(node);
 	node = NULL;
